M0 T310 S8 retain COMMAND native lengths

+++ src/mvdm-host/dos/command/cmdmisc.c
@@ -42,7 +42,10 @@ VOID cmdGetNextCmd (VOID)
 {
 LPSTR   lpszCmd;
 PCMDINFO pCMDInfo;
-ULONG   cb;
+/* DIVERGENCE(MVDM-HOST-DIV-115): retain local title and command-line
+ * lengths at their native width.  The existing guest byte write remains at
+ * the source's already asserted 127-byte boundary. */
+size_t  cb;
 PREDIRCOMPLETE_INFO pRdrInfo;
 VDMINFO MyVDMInfo;
 
@@ -130,7 +133,7 @@ char    CmdLine[MAX_PATH];
 		    cb = strlen(achTitle);
 		    // GetConsoleTitleA and SetConsoleTitleA
 		    // are working on OEM character set.
-		    GetConsoleTitleA(achTitle + cb, MAX_PATH - cb - 1);
+		    GetConsoleTitleA(achTitle + cb, (DWORD)(MAX_PATH - cb - 1));
 		    cb = strlen(achTitle);
 		    achTitle[cb] = ']';
 		    achTitle[cb + 1] = '\0';
